Maint: move dynatrace out of cluster command and update sort order
